+++ robotframework-ls/src/robotframework_ls/impl/collect_keywords.py
@@ -157,6 +157,7 @@ class _KeywordFoundFromAst(object):
 class _KeywordFoundFromLibrary(object):
 
     __slots__ = [
+        "_lib_deprecated",
         "_library_doc",
         "_library_alias",
         "_keyword_doc",
@@ -169,6 +170,7 @@ class _KeywordFoundFromLibrary(object):
 
     def __init__(
         self,
+        lib_deprecated,
         library_doc,
         keyword_doc,
         keyword_name,
@@ -177,7 +179,7 @@ class _KeywordFoundFromLibrary(object):
         completion_item_kind,
         library_alias=None,
     ):
-
+        self._lib_deprecated = lib_deprecated
         self._library_doc = library_doc
         self._keyword_doc = keyword_doc
         self._keyword_name = keyword_name
@@ -213,6 +215,8 @@ class _KeywordFoundFromLibrary(object):
 
     @instance_cache
     def is_deprecated(self):
+        if self._lib_deprecated:
+            return self._lib_deprecated
         from robotframework_ls.impl import text_utilities
 
         return text_utilities.has_deprecated_text(self._keyword_doc.doc)
@@ -333,6 +337,7 @@ def _collect_libraries_keywords(
 ):
     from robotframework_ls.impl.libspec_manager import LibspecManager
     from robotframework_ls.impl.protocols import ILibraryDocOrError
+    from robotframework_ls.impl import text_utilities
 
     # Get keywords from libraries
     from robocorp_ls_core.lsp import CompletionItemKind
@@ -373,6 +378,9 @@ def _collect_libraries_keywords(
             else:
                 memo[key] = True
 
+            doc = library_doc.doc
+            lib_deprecated = doc and text_utilities.has_deprecated_text(doc)
+
             #: :type keyword: KeywordDoc
             for keyword in library_doc.keywords:
                 keyword_name = keyword.name
@@ -391,6 +399,7 @@ def _collect_libraries_keywords(
 
                     collector.on_keyword(
                         _KeywordFoundFromLibrary(
+                            lib_deprecated,
                             library_doc,
                             keyword,
                             keyword_name,